+++ lib/shadcn_ui.ex
@@ -25,6 +25,7 @@ defmodule ShadcnUI do
     ShadcnUI.Components.Overlays.Popover,
     ShadcnUI.Components.Overlays.DropdownActions,
     ShadcnUI.Components.Overlays.Tooltip,
+    ShadcnUI.Components.Overlays.HoverCard,
     ShadcnUI.Components.Foundation.Alert,
     ShadcnUI.Components.Foundation.Avatar,
     ShadcnUI.Components.Foundation.Badge,